requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
distro==1.9.0
exceptiongroup==1.3.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mss==10.1.0
packaging==25.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
PyQt6==6.10.2
PyQt6-Qt6==6.10.1
PyQt6_sip==13.11.0
pytesseract==0.3.13
python-dotenv==1.2.1
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
